# frontend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/system/api-keys/{id}": {
      "patch": {
        "operationId": "SystemApiKeysController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SystemApiKeys"
        ]
      },
      "delete": {
        "operationId": "SystemApiKeysController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SystemApiKeys"
        ]
      }
    },
...
    "/api/system/apis/{id}": {
      "patch": {
        "operationId": "SystemApisController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApiPermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SystemApis"
        ]
      },
      "delete": {
        "operationId": "SystemApisController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SystemApis"
        ]
      }
    },
...
    "/api/system/integrations/{id}": {
      "patch": {
        "operationId": "SystemIntegrationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SystemIntegrations"
        ]
      },
      "delete": {
        "operationId": "SystemIntegrationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "SystemIntegrations"
        ]
      }
    },
...
  },
  "info": {
    "title": "AiSystem API",
    "description": "AiSystem 后端接口文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateApiPermissionDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}